modified viz module header layout for 2.4 compatibility
